Context resetting when entering safe areas
In order to avoid enemies not spawning, the delivery of
S2CContextSetContextBaseNtc was commented out.
However this had the side effect of human enemies not spawning at all.
To have the best of both worlds, we send S2CInstance_13_42_16_Ntc, as
it's been noticed on the pcaps, when entering a safe area (White Dragon
Temple) so the context gets reset and enemies in the field respawn.