Updated some values and fix an issue
Fixed an issue where it was giving an extra level status when leveling to level 2.
Now the only issues remaining is that it won't update status if the leveling difference is too high, like leveling 25+ levels at once. The other issue is that the status should be using double and not ushort, that means they are being converted and not updating into the db.