feat: Add support for GP course effects
Added support for GP course effects which are configurable by the new
file GpCourseInfo.json. Extracted 15 course entries from the packet
captures and encoded them into this file. Also added a new custom course
called "Arrowgene QOL Course" (quality of life). This is an example of
how to create new course effects which were not present when the game
was live.