feat: Track HP and Stamina properly after upgrades and logout
When a user logs out or disconnects, the disconnect notification is raised.
When this happens, the current value of green and white HP is stored in the
DB. This value will then be used the next time the character is logged into.
When the user purchases an upgrade at the white dragon, HP and Stamina are
refilled to the max value.