Implement database versioning and migrations
Added support for fluentmigrator and translated the project schema file
into a migration file. Added support for SQLite, MariaDB and Postgres.
When the server first runs, it will check to see if any migrations are
required and applies them.

Added a new command 'dbmigration' to the cli tool. This can be used to
downgrade or upgrade the database without directly starting the server.