Periodically check age of client pings
To prevent half-connected sockets, both the login and game server's ping handlers
now use a timer to periodically check if the last ping has been too long ago.
If it has been, the server will try sending a packet to the client, if the packet goes through, the client is still alive
and no further action is needed;
if the packet doesn't go through, the client is dead, the write will fail, and it'll be cleaned up.

This will (hopefully) solve the S-6008 error, where the server thinks the user is still connected to the server and therefore prevents new connections from them