fix: Fix some oddities with character
Attempt to fix an issue where the player sometimes is "dead" when they
log in, join parties or switch to some hub area. It seems that the HP is
being saved as 0 sometimes in the DB. So when the context gets updated
the player is dead, unable to select any menu options.