Fix the white dragon in cutscenes
Seems the 2.0 cutscenes include the white dragon in the animation and
doesn't remove the one sitting in the chamber by itself. Since this
happens, you end up with 2 white dragons at different points in their
animation loop. In cutscenes which include the white dragon in the
audience chamber, we need to despawn the dragon until the cutscene is
over and then respawn it after.