refactor: Change #load root in scripts
Change the #load path to be the script root directory. This makes it
easier to support vscode autocompletion for everyone with less setup.
It also forces scripts to apprend the "lib/" prefix when loading
libraries making it more clear where the file comes from. Added a new
libs file to help with auto completions being easier to setup. Updated
the README.md in the root script directory with directions.